fix(dockerfile): Sentry integration (#153)
* build(main): Log Sentry configuration at startup

This commit introduces console logs to display the Sentry DSN, environment, and release version upon application startup. This information can be helpful for debugging and verifying the Sentry configuration in different environments.

* fix(dockerfile): Sentry integration

This commit fixes Sentry integration by:
- Adding `SENTRY_AUTH_TOKEN` and `VITE_SENTRY_DSN` build arguments.
- Setting the `VITE_SENTRY_DSN` environment variable.

This allows the application to report errors and monitor performance using Sentry.